macos: become Phantom.app, style panel in Appearance, behaviors tab
The product renames to Phantom: PRODUCT_NAME on the app target (dock
tile plugin untouched), the zig packaging path follows, and
CFBundleName no longer needs post-processing — the Dock and menu bar
agree on the name. Every style control (terminal font/cursor, window
background/opacity/blur, sidebar background/width/divider, tab item
density) moves into an Appearance style panel sectioned by area,
leaving the General and Sidebar tabs behavior-only, plus a new
Behaviors tab (window restore, agent resume, new-terminal position)
grouped by area. The sidebar divider becomes configurable (default,
hidden or custom color). The regular glass background keeps the
theme tint over the glass; only the clear variant is colorless.
