+++ macos/Sources/Features/Settings/SettingsView.swift
@@ -1,7 +1,8 @@
 import SwiftUI
 
 /// Root of the settings window: section list on the left, the selected
-/// section's form on the right.
+/// section's form on the right. Appearance owns every style control;
+/// the other sections hold behavior only.
 struct SettingsRootView: View {
     let ghostty: Ghostty.App
 
@@ -11,6 +12,7 @@ struct SettingsRootView: View {
         case general
         case appearance
         case sidebar
+        case behaviors
 
         var id: String { rawValue }
 
@@ -19,6 +21,7 @@ struct SettingsRootView: View {
             case .general: return "General"
             case .appearance: return "Appearance"
             case .sidebar: return "Sidebar"
+            case .behaviors: return "Behaviors"
             }
         }
 
@@ -27,6 +30,7 @@ struct SettingsRootView: View {
             case .general: return "gearshape"
             case .appearance: return "paintpalette"
             case .sidebar: return "sidebar.left"
+            case .behaviors: return "slider.horizontal.3"
             }
         }
     }
@@ -49,130 +53,21 @@ struct SettingsRootView: View {
                 AppearanceSettingsView(ghostty: ghostty, store: store)
             case .sidebar:
                 SidebarSettingsView(ghostty: ghostty, store: store)
+            case .behaviors:
+                BehaviorsSettingsView(ghostty: ghostty, store: store)
             }
         }
         .frame(minWidth: 700, minHeight: 480)
     }
 }
 
-/// Basic terminal options: font, opacity, cursor.
+/// General behavior: access to the raw configuration.
 struct GeneralSettingsView: View {
     let ghostty: Ghostty.App
     @ObservedObject var store: GuiConfigStore
 
-    @State private var fontFamily: String = ""
-    @State private var fontSize: Double = 13
-    @State private var backgroundOpacity: Double = 1
-    @State private var blurMode: String = "off"
-    @State private var blurRadius: Double = 20
-    @State private var backgroundColorOverride: Color?
-    @State private var cursorStyle: String = ""
-
-    private static let cursorStyles: [(value: String, label: String)] = [
-        ("", "Default"),
-        ("block", "Block"),
-        ("bar", "Bar"),
-        ("underline", "Underline"),
-        ("block_hollow", "Hollow Block"),
-    ]
-
     var body: some View {
         Form {
-            Section("Font") {
-                LabeledContent("Family") {
-                    TextField("", text: $fontFamily, prompt: Text("System default"))
-                        .labelsHidden()
-                        .multilineTextAlignment(.leading)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .onSubmit { apply("font-family", fontFamily) }
-                }
-
-                LabeledContent("Size") {
-                    HStack {
-                        Slider(value: $fontSize, in: 8...32, step: 1) { editing in
-                            if !editing { apply("font-size", String(Int(fontSize))) }
-                        }
-                        Text("\(Int(fontSize)) pt")
-                            .monospacedDigit()
-                            .foregroundStyle(.secondary)
-                            .frame(width: 44, alignment: .trailing)
-                    }
-                }
-            }
-
-            Section("Window") {
-                LabeledContent("Background Color") {
-                    HStack(spacing: 8) {
-                        if backgroundColorOverride != nil {
-                            Button("Use Theme Color") {
-                                backgroundColorOverride = nil
-                                apply("background", "")
-                            }
-                            .buttonStyle(.link)
-                            .font(.caption)
-                        }
-
-                        ColorPicker(
-                            "",
-                            selection: Binding(
-                                get: { backgroundColorOverride ?? .black },
-                                set: { newValue in
-                                    backgroundColorOverride = newValue
-                                    apply("background", NSColor(newValue).hexString ?? "")
-                                }
-                            ),
-                            supportsOpacity: false
-                        )
-                        .labelsHidden()
-                    }
-                }
-
-                LabeledContent("Background Opacity") {
-                    HStack {
-                        Slider(value: $backgroundOpacity, in: 0.3...1) { editing in
-                            if !editing {
-                                apply("background-opacity", formatOpacity(backgroundOpacity))
-                            }
-                        }
-                        Text(formatOpacity(backgroundOpacity))
-                            .monospacedDigit()
-                            .foregroundStyle(.secondary)
-                            .frame(width: 44, alignment: .trailing)
-                    }
-                }
-
-                Picker("Background Blur", selection: $blurMode) {
-                    Text("Off").tag("off")
-                    Text("Blur Radius").tag("radius")
-                    Text("Glass").tag("glass-regular")
-                    Text("Glass (Clear)").tag("glass-clear")
-                }
-                .onChange(of: blurMode) { _ in applyBlur() }
-
-                if blurMode == "radius" {
-                    LabeledContent("Blur Intensity") {
-                        HStack {
-                            Slider(value: $blurRadius, in: 1...80, step: 1) { editing in
-                                if !editing { applyBlur() }
-                            }
-                            Text("\(Int(blurRadius))")
-                                .monospacedDigit()
-                                .foregroundStyle(.secondary)
-                                .frame(width: 44, alignment: .trailing)
-                        }
-                    }
-                }
-
-                Picker("Cursor Style", selection: $cursorStyle) {
-                    ForEach(Self.cursorStyles, id: \.value) { style in
-                        Text(style.label).tag(style.value)
-                    }
-                }
-                .onChange(of: cursorStyle) { value in
-                    apply("cursor-style", value)
-                }
-            }
-
             Section {
                 LabeledContent("Config File") {
                     Button("Open in Editor") {
@@ -180,84 +75,27 @@ struct GeneralSettingsView: View {
                     }
                 }
             } footer: {
-                Text("Settings changed here are stored in \(GuiConfigStore.fileName) and included from your config file. Hand-written options stay untouched.")
+                Text("Settings changed in this window are stored in \(GuiConfigStore.fileName) and included from your config file. Hand-written options stay untouched. Style options (fonts, colors, blur) live in Appearance.")
                     .font(.caption)
                     .foregroundStyle(.secondary)
             }
         }
         .formStyle(.grouped)
         .navigationTitle("General")
-        .onAppear { populate() }
-    }
-
-    private func populate() {
-        fontFamily = store.string("font-family") ?? ""
-        fontSize = store.double("font-size", default: 13)
-        backgroundOpacity = store.double("background-opacity", default: 1)
-        switch store.string("background-blur") ?? "false" {
-        case "false":
-            blurMode = "off"
-        case "true":
-            blurMode = "radius"
-            blurRadius = 20
-        case "macos-glass-regular":
-            blurMode = "glass-regular"
-        case "macos-glass-clear":
-            blurMode = "glass-clear"
-        case let raw:
-            if let value = Double(raw), value > 0 {
-                blurMode = "radius"
-                blurRadius = value
-            } else {
-                blurMode = "off"
-            }
-        }
-        backgroundColorOverride = store.string("background")
-            .flatMap { NSColor(hex: $0) }
-            .map { Color(nsColor: $0) }
-        cursorStyle = store.string("cursor-style") ?? ""
-    }
-
-    private func apply(_ key: String, _ value: String) {
-        store.set(key, value.isEmpty ? nil : value)
-        store.apply(ghostty: ghostty)
-    }
-
-    private func applyBlur() {
-        let value: String
-        switch blurMode {
-        case "radius": value = String(Int(blurRadius))
-        case "glass-regular": value = "macos-glass-regular"
-        case "glass-clear": value = "macos-glass-clear"
-        default: value = "false"
-        }
-        apply("background-blur", value)
-    }
-
-    private func formatOpacity(_ value: Double) -> String {
-        String(format: "%.2f", value)
     }
 }
 
-/// Options for the tab sidebar itself.
+/// Sidebar behavior: visibility, ordering and which tab info shows.
 struct SidebarSettingsView: View {
     let ghostty: Ghostty.App
     @ObservedObject var store: GuiConfigStore
 
     @State private var sidebarEnabled: Bool = false
-    @State private var sidebarWidth: Double = 240
 
     @AppStorage("SidebarShowDirectory") private var showDirectory = true
     @AppStorage("SidebarShowGitBranch") private var showGitBranch = true
     @AppStorage("SidebarShowGitStatus") private var showGitStatus = true
     @AppStorage("SidebarShowPullRequest") private var showPullRequest = true
-    @AppStorage("SidebarRestoreAgentSessions") private var restoreAgentSessions = true
-    @AppStorage("SidebarNewTabPosition") private var newTabPosition = "end"
-    @AppStorage("SidebarTabDensity") private var tabDensity = "default"
-
-    @State private var backgroundMode: String = "theme"
-    @State private var tintColor: Color = .black
-    @State private var tintOpacity: Double = 0
 
     var body: some View {
         Form {
@@ -269,71 +107,13 @@ struct SidebarSettingsView: View {
                         store.apply(ghostty: ghostty)
                     }
 
-                LabeledContent("Default Width") {
-                    HStack {
-                        Slider(value: $sidebarWidth, in: 180...480, step: 10) { editing in
-                            if !editing {
-                                store.set("sidebar-width", String(Int(sidebarWidth)))
-                                store.apply(ghostty: ghostty)
-                            }
-                        }
-                        Text("\(Int(sidebarWidth)) pt")
-                            .monospacedDigit()
-                            .foregroundStyle(.secondary)
-                            .frame(width: 48, alignment: .trailing)
-                    }
-                }
             } footer: {
-                Text("The sidebar toggle applies to new windows. Dragging the divider overrides the default width.")
+                Text("The sidebar toggle applies to new windows. Sidebar style (background, width, tab item look) lives in Appearance.")
                     .font(.caption)
                     .foregroundStyle(.secondary)
             }
 
-            Section {
-                Picker("New Terminal Position", selection: $newTabPosition) {
-                    Text("Bottom of List").tag("end")
-                    Text("Top of List").tag("start")
-                }
-            }
-
-            Section {
-                Picker("Background", selection: $backgroundMode) {
-                    Text("Match Theme").tag("theme")
-                    Text("Match Window").tag("window")
-                    Text("Custom Tint").tag("custom")
-                }
-                .onChange(of: backgroundMode) { _ in saveTint() }
-
-                if backgroundMode == "custom" {
-                    LabeledContent("Tint") {
-                        HStack {
-                            Slider(value: $tintOpacity, in: 0...1) { editing in
-                                if !editing { saveTint() }
-                            }
-
-                            ColorPicker("", selection: $tintColor, supportsOpacity: false)
-                                .labelsHidden()
-                                .onChange(of: tintColor) { _ in saveTint() }
-
-                            Text(tintOpacity <= 0.001 ? "Off" : String(format: "%.0f%%", tintOpacity * 100))
-                                .monospacedDigit()
-                                .foregroundStyle(.secondary)
-                                .frame(width: 40, alignment: .trailing)
-                        }
-                    }
-                }
-            } footer: {
-                Text("Backgrounds are layered: the window background (General) is the base for every pane and the terminal adds its theme background. Match Theme keeps the sidebar in the theme's color, Match Window leaves only the base showing, and Custom Tint sets the sidebar apart with its own color.")
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-            }
-
-            Section("Tab Item") {
-                Picker("Style", selection: $tabDensity) {
-                    Text("Default").tag("default")
-                    Text("Compact").tag("compact")
-                }
-
+            Section("Tab Info") {
                 Toggle("Show Working Directory", isOn: $showDirectory)
                     .toggleStyle(.switch)
                 Toggle("Show Git Branch", isOn: $showGitBranch)
@@ -344,39 +124,59 @@ struct SidebarSettingsView: View {
                     .toggleStyle(.switch)
             }
 
-            Section {
-                Toggle("Resume Agent Sessions on Restore", isOn: $restoreAgentSessions)
-                    .toggleStyle(.switch)
-            } footer: {
-                Text("When windows are restored, tabs that were running a Claude Code session run `claude --continue` to pick the conversation back up.")
-                    .font(.caption)
-                    .foregroundStyle(.secondary)
-            }
         }
         .formStyle(.grouped)
         .navigationTitle("Sidebar")
         .onAppear {
             sidebarEnabled = store.bool("sidebar")
-            sidebarWidth = store.double("sidebar-width", default: 240)
-
-            let defaults = UserDefaults.standard
-            backgroundMode = defaults.string(forKey: "SidebarBackgroundMode") ?? "theme"
-            tintOpacity = defaults.double(forKey: "SidebarTintOpacity")
-            if let hex = defaults.string(forKey: "SidebarTintHex"),
-               let color = NSColor(hex: hex) {
-                tintColor = Color(nsColor: color)
-            }
         }
     }
+}
 
-    private func saveTint() {
-        let defaults = UserDefaults.standard
-        defaults.set(backgroundMode, forKey: "SidebarBackgroundMode")
-        defaults.set(NSColor(tintColor).hexString ?? "#000000", forKey: "SidebarTintHex")
-        defaults.set(tintOpacity, forKey: "SidebarTintOpacity")
-        NotificationCenter.default.post(
-            name: TerminalController.sidebarTintDidChange,
-            object: nil
-        )
+
+/// Behavioral options grouped by area — nothing here changes looks.
+struct BehaviorsSettingsView: View {
+    let ghostty: Ghostty.App
+    @ObservedObject var store: GuiConfigStore
+
+    @AppStorage("SidebarRestoreAgentSessions") private var restoreAgentSessions = true
+    @AppStorage("SidebarNewTabPosition") private var newTabPosition = "end"
+
+    @State private var restoreWindows = true
+
+    var body: some View {
+        Form {
+            Section("General") {
+                Toggle("Restore Windows on Launch", isOn: $restoreWindows)
+                    .toggleStyle(.switch)
+                    .onChange(of: restoreWindows) { value in
+                        store.set("window-save-state", value ? "always" : "default")
+                        store.apply(ghostty: ghostty)
+                    }
+            }
+
+            Section {
+                Toggle("Resume Agent Sessions on Restore", isOn: $restoreAgentSessions)
+                    .toggleStyle(.switch)
+            } header: {
+                Text("Terminal")
+            } footer: {
+                Text("When windows are restored, tabs that were running a Claude Code session run `claude --continue` to pick the conversation back up.")
+                    .font(.caption)
+                    .foregroundStyle(.secondary)
+            }
+
+            Section("Sidebar") {
+                Picker("New Terminal Position", selection: $newTabPosition) {
+                    Text("Bottom of List").tag("end")
+                    Text("Top of List").tag("start")
+                }
+            }
+        }
+        .formStyle(.grouped)
+        .navigationTitle("Behaviors")
+        .onAppear {
+            restoreWindows = (store.string("window-save-state") ?? "always") == "always"
+        }
     }
 }
