+++ macos/Sources/Features/Terminal/Sidebar/SidebarLayoutModel.swift
@@ -1,5 +1,6 @@
-import Foundation
+import AppKit
 import Combine
+import Foundation
 
 /// App-wide sidebar collapse state. Shared by every window so switching
 /// tabs never changes the sidebar geometry, and persisted so it
@@ -25,4 +26,24 @@ final class SidebarCollapseState: ObservableObject {
 final class SidebarLayoutModel: ObservableObject {
     /// Creates a new terminal tab in this window's tab group.
     var onNewTab: () -> Void = {}
+}
+
+/// The sidebar | terminal split view, with a user-configurable divider:
+/// default system color, hidden, or a custom color.
+final class SidebarSplitView: NSSplitView {
+    override var dividerColor: NSColor {
+        let defaults = UserDefaults.standard
+        switch defaults.string(forKey: "SidebarDividerMode") ?? "default" {
+        case "hidden":
+            return .clear
+        case "custom":
+            if let hex = defaults.string(forKey: "SidebarDividerColorHex"),
+               let color = NSColor(hex: hex) {
+                return color
+            }
+            return super.dividerColor
+        default:
+            return super.dividerColor
+        }
+    }
 }
